+++ tests/unit/anchore_engine/services/policy_engine/engine/vulns/test_mappers.py
@@ -23,6 +23,7 @@ from anchore_engine.services.policy_engine.engine.vulns.mappers import (
         pytest.param("alpine", "alpine", "alpine", id="alpine"),
         pytest.param("sles", "sles", "sles", id="sles"),
         pytest.param("windows", "windows", "", id="windows"),
+        pytest.param("rocky", "rockylinux", "fedora", id="rocky"),
     ],
 )
 def test_engine_distro_mappers(test_distro, expected_os, expected_like_os):